인증 로직 관련 util 변경

+++ backend/src/main/java/com/joo/abysshop/controller/cart/CartItemController.java
@@ -4,9 +4,12 @@ import com.joo.abysshop.dto.cart.request.AddItemToCartRequest;
 import com.joo.abysshop.dto.cart.request.DeleteItemFromCartRequest;
 import com.joo.abysshop.dto.cart.request.UpdateCartItemsQuantityRequest;
 import com.joo.abysshop.service.cart.CartItemCommandService;
+import com.joo.abysshop.util.security.CustomUserDetails;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -23,21 +26,36 @@ public class CartItemController {
 
     @PostMapping("/add")
     public ResponseEntity<Void> addItemToCart(
-        @RequestBody AddItemToCartRequest addItemToCartRequest) {
+        @RequestBody AddItemToCartRequest addItemToCartRequest,
+        @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        if (customUserDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         cartItemCommandService.addOrUpdateCartItem(addItemToCartRequest);
         return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/delete")
     public ResponseEntity<Void> deleteItemFromCart(
-        @RequestBody DeleteItemFromCartRequest deleteItemFromCartRequest) {
+        @RequestBody DeleteItemFromCartRequest deleteItemFromCartRequest,
+        @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        if (customUserDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         cartItemCommandService.deleteCartItem(deleteItemFromCartRequest);
         return ResponseEntity.noContent().build();
     }
 
     @PatchMapping("/quantity")
     public ResponseEntity<Void> updateCartItemsQuantity(
-        @RequestBody List<UpdateCartItemsQuantityRequest> requestList) {
+        @RequestBody List<UpdateCartItemsQuantityRequest> requestList,
+        @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        if (customUserDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         cartItemCommandService.updateCartItemsQuantity(requestList);
         return ResponseEntity.noContent().build();
     }
